+++ changelog.md
@@ -362,3 +362,6 @@
     // ver 0.7.3 @ 2023-10-18
     //    jw: fixed logic of not prefering manual visit of imdb (of v0.6.31)
     //    jw: dirty fix for the case of 'Higurashi no naku koro ni'
+    // ver 0.7.5 @ 2023-10-29
+    //    jw: fixed logic when result is none
+    //    app & setting: preparing to refactor not to use numberToBaseEl, targetEl, determinePathnameBy, etc.